refactor(buildrequest): untag parseCallUnsupportedProvidersEnv
The function operates on os.Getenv + strings.Split + strings.TrimSpace
— pure stdlib with no dependency on the //go:build debug-tagged state.
Move it to a new untagged call_support_parse.go; the rest of
call_support_debug.go (the cached/Once-gated production seam,
debugFilterCallSupported) stays tagged alongside the call_support_stub.go
no-op partner.

Rename call_support_debug_test.go to call_support_parse_test.go and
drop the //go:build debug line. There is no -tags=debug step in
unit-tests.yml today, so TestParseCallUnsupportedProvidersEnv was
silently skipped on every CI run; this commit moves it into the
default per-module test loop. Pure-additive coverage gain.

Regenerate //go:embed source lists in bamlutils/embed.go (picks up the
new call_support_parse.go) and the top-level embed.go (picks up
cmd/serve/unary_handlers.go from commit 855d358).
