+++ bamlutils/buildrequest/call_support_parse_test.go
@@ -1,5 +1,3 @@
-//go:build debug
-
 package buildrequest
 
 import (
